pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-job-radar"
version = "0.2.0"
description = "Personal LLM-powered job-matching radar with Telegram notifications."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Roberto Mena Rodríguez" }]
keywords = ["llm", "gemini", "rag", "jobs", "telegram", "automation"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Office/Business",
]
dependencies = [
    "feedparser>=6.0.11",
    "requests>=2.32.3",
    "python-dotenv>=1.0.1",
    "google-generativeai>=0.8.3",
    "beautifulsoup4>=4.12.3",
    "lxml>=5.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.6",
]

[project.scripts]
ai-job-radar = "ai_job_radar.cli:main"

[project.urls]
Homepage = "https://github.com/RobertoMena/ai-job-radar"
Issues = "https://github.com/RobertoMena/ai-job-radar/issues"

[tool.setuptools.packages.find]
include = ["ai_job_radar*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]  # line length handled by formatter

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q --strict-markers"
